Fix merge conflict residue in filter and browser_input
- filter/main.rs: remove spurious as f64 cast; both operands are i32 as in master
- browser_input/input.rs: add missing space before ] in CLIENT and PAGE closures

+++ examples/minwebgl/filter/src/main.rs
@@ -72,8 +72,8 @@ fn run()
         let rect = canvas.get_bounding_client_rect();
         let canvas_x = rect.left() as i32;
         let canvas_y = rect.top() as i32;
-        let x = ( e.client_x() - canvas_x as f64 ) as f32;
-        let y = ( e.client_y() - canvas_y as f64 ) as f32;
+        let x = ( e.client_x() - canvas_x ) as f32;
+        let y = ( e.client_y() - canvas_y ) as f32;
         let y = canvas.height() as f32 - y;
         gl::uniform::upload( &gl, cursor_pos_location.clone(), [ x, y ].as_slice() ).unwrap();
         gl.draw_arrays( GL::TRIANGLES, 0, 3 );
